% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arti_fr.R
\name{calculAgregatsZones}
\alias{calculAgregatsZones}
\title{Fonction calculant les agrégats sur une zone à partir des carreaux qui la recouvrent
https://cerema.app.box.com/v/pnb-action7-indicateurs-ff/file/1270338746048
fonction inspirée de celle du CEREMA proposée sur https://www.insee.fr/fr/statistiques/4176290?sommaire=4176305}
\usage{
calculAgregatsZones(fichierContoursSHP, listeCodesCommune)
}
\arguments{
\item{listeCodesCommune}{Liste des codes Insee des communes}

\item{cheminFichierContoursSHP}{Chemin des contours de la ou des zones à croiser avec les carreaux de la bdd filosofi}
}
\value{
Une table contenant une ligne pour chaque zone du fichier de contours
}
\description{
Fonction calculant les agrégats sur une zone à partir des carreaux qui la recouvrent
https://cerema.app.box.com/v/pnb-action7-indicateurs-ff/file/1270338746048
fonction inspirée de celle du CEREMA proposée sur https://www.insee.fr/fr/statistiques/4176290?sommaire=4176305
}
\examples{
calculAgregatsZones("c:/mesdocs/moncontour.shp", c("01001","01002"))

}
