% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arti_fr.R
\name{carto_conso_espace_dept_fr}
\alias{carto_conso_espace_dept_fr}
\title{Carte de la consommation d'espaces par département en France avec
Pour chaque département :
  - le total de la surface artificialisée
  - la part de la surface totale qui a été artificialisée entre 2009 et 2022
  - la part de la surface artificialisée lié à l’activité et à l’habitat
Chargement des Indicateurs de consommation d'espace en 2022
TODO traiter les 5 lignes ci-dessus}
\usage{
carto_conso_espace_dept_fr()
}
\value{
une carte
}
\description{
Carte de la consommation d'espaces par département en France avec
Pour chaque département :
  - le total de la surface artificialisée
  - la part de la surface totale qui a été artificialisée entre 2009 et 2022
  - la part de la surface artificialisée lié à l’activité et à l’habitat
Chargement des Indicateurs de consommation d'espace en 2022
TODO traiter les 5 lignes ci-dessus
}
