% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arti_fr.R
\name{carto_conso_espace_oree_anjou}
\alias{carto_conso_espace_oree_anjou}
\title{sur Orée d'Anjou
Carroyage sur 200m de la consommation d'espaces}
\usage{
carto_conso_espace_oree_anjou()
}
\value{
une carte tmap
}
\description{
sur Orée d'Anjou
Carroyage sur 200m de la consommation d'espaces
}
