% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arti_fr.R
\name{histo_conso_espace_oree_anjou}
\alias{histo_conso_espace_oree_anjou}
\title{sur l'ECPI ou la commune sélectionnée entre 2009 et 2024
Consommation d'espaces :
Conversion par 6 types d'activité en hectares avec 2 décimales
de surface non artificialisée en surface artificialisée
soit un histogrammes avec répartition de part d'activité dans chacunes des colonnes}
\usage{
histo_conso_espace_oree_anjou()
}
\value{
un histogramme
}
\description{
sur l'ECPI ou la commune sélectionnée entre 2009 et 2024
Consommation d'espaces :
Conversion par 6 types d'activité en hectares avec 2 décimales
de surface non artificialisée en surface artificialisée
soit un histogrammes avec répartition de part d'activité dans chacunes des colonnes
}
