% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arti_fr.R
\name{histo_conso_espace_varenne}
\alias{histo_conso_espace_varenne}
\title{Evolution de la consommation d'espaces sur la commune déléguée de La Varenne (49270)}
\usage{
histo_conso_espace_varenne()
}
\value{
un histogramme
}
\description{
Evolution de la consommation d'espaces sur la commune déléguée de La Varenne (49270)
}
