% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_artificialisation_st_herblain.R
\name{mod_arti_44_ui}
\alias{mod_arti_44_ui}
\alias{mod_arti_44_server}
\title{mod_arti_44_ui and mod_arti_44_server}
\usage{
mod_arti_44_ui(id)

mod_arti_44_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
