% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_prix_foncier_nantes.R,
%   R/mod_04_prix_foncier_angers.R
\name{mod_foncier_ui}
\alias{mod_foncier_ui}
\alias{mod_foncier_server}
\title{mod_foncier_ui and mod_foncier_server}
\usage{
mod_foncier_ui(id)

mod_foncier_server(input, output, session)

mod_foncier_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
