% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arti_fr.R
\name{histo_conso_espace_france}
\alias{histo_conso_espace_france}
\title{Evolution de la consommation d'espaces par région en France}
\usage{
histo_conso_espace_france()
}
\value{
un graphique en courbe
}
\description{
Evolution de la consommation d'espaces par région en France
}
